input fixtures are now just created anonymously
