--- tests/ValueResolver/Fixture/EmptyCommand.php
@@ -1,9 +1,0 @@
-<?php
-
-namespace OneToMany\RichBundle\Tests\ValueResolver\Fixture;
-
-use OneToMany\RichBundle\Contract\CommandInterface;
-
-final readonly class EmptyCommand implements CommandInterface
-{
-}
